requirements.txt
joblib==1.5.1
matplotlib==3.10.3
numpy==2.3.0
pandas==2.3.0
tqdm==4.67.1
tornado==6.5.1
seaborn==0.13.2
openpyxl==3.1.5
pyyaml==6.0.2